+++ NEWS.md
@@ -1,6 +1,8 @@
 # patRoon 0.1.0.9000
 
 ## Decomber 2018
+* Minimum intensity arguments for Bruker DataAnalysis MS peak lists.
+* Slightly different `doFMF` beahviour for DataAnalysis feature finding.
 * Several important updates were made to fomula calculation functionality.
     * The interface has been simplified as the functionality from the `formula` and `formulaConsensus` classes are now merged: there is no need to call `consensus()` anymore after `generateFormulas()`.
     * Formulae can now directly be calculated for feature groups by using group averaged MS peak lists (by setting `calculateFeatures=FALSE`). This can greatly speed up calulcation, especially with many analyses.